requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
pandas
numpy
openpyxl
matplotlib
seaborn
tabulate
langchain
langchain-experimental
langchain-groq
python-dotenv